+++ extensions/cornerstone/src/services/SyncGroupService/createHydrateSegmentationSynchronizer.ts
@@ -65,13 +65,11 @@ const segmentationRepresentationModifiedCallback = async (
     return;
   }
 
-  // whatever type the source viewport has, we need to add that to the target viewport
-  const sourceViewportRepresentation = segmentationService.getSegmentationRepresentations(
-    sourceViewport.viewportId,
-    { segmentationId }
-  );
-
-  const type = sourceViewportRepresentation[0].type;
+  // Ensure the segmentation representation aligns with the target viewport type.
+  const type: Enums.SegmentationRepresentations =
+    viewport.type === CoreEnums.ViewportType.VOLUME_3D
+      ? Enums.SegmentationRepresentations.Surface
+      : Enums.SegmentationRepresentations.Labelmap;
 
   await segmentationService.addSegmentationRepresentation(targetViewportId, {
     segmentationId,